feat(frontend): logout user if auth cookie invalid
